A GruntJS plugin that runs BusterJS tests
This initial implementation of the GruntJS plugin adds support for starting a BusterJS server, a PhantomJS instance, running BusterJS tests, shutting down the PhantomJS instance, and shutting down the BusterJS server.

The configuration file to be used, configuration groups to test, server port, and output reporter can all be configured.