Break monolithic functionality into multiple modules
To support the change to a multi-task, the functionality of the plugin has been broken out into multiple modules.  These modules include a factory for child processes, utilities for dealing with configuration, a finder for executables, a PhantomJS lifecycle manager, a BusterJS server lifecycle manager, and a BusterJS test runner.  In addition to the implementations now being broken out, they are all completely tested.

At this point, the original task has not been changed to use the new modules.  This plugin is a special case in that it uses its own code to test itself and therefore the task needs to stay intact for one more commit.